requirements.txt
# Core dependencies
httpx>=0.24.0
prometheus-client>=0.16.0
colorama>=0.4.6

# A2A SDK (required for native executor interface)
a2a-server>=0.2.0

# Server dependencies (v1.0.0)
uvicorn>=0.22.0
starlette>=0.27.0

# Optional system monitoring
psutil>=5.9.0
pynvml>=11.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0

acp-sdk >= 1.0.0

# Safety Tech scenario dependencies
kagglehub>=0.2.0
datasets>=2.14.0
pandas>=2.0.0
PyYAML>=6.0.0

